feat(accounting): implement custom hook for accounting records, refactor api restructure accounting record interfaces
