+++ src/app/dashboard/accounting-book/Records.tsx
@@ -1,24 +1,5 @@
+import { useAccountingRecords } from '@/hooks/useAccountingRecords';
 import { getCategoryIcon, getCategoryLabel } from '@/lib/categories';
-import { db } from '@/lib/firebase';
-import { endOfMonth, isSameDay, startOfMonth } from 'date-fns';
-import {
-  collection,
-  onSnapshot,
-  query,
-  Timestamp,
-  where,
-} from 'firebase/firestore';
-import { useEffect, useState } from 'react';
-
-interface AccountingRecord {
-  id: string;
-  date: Timestamp;
-  amount: string;
-  category: string;
-  account: string;
-  note?: string;
-  images?: string[];
-}
 
 interface RecordsProps {
   date: Date | undefined;
@@ -26,77 +7,7 @@ interface RecordsProps {
 }
 
 export default function Records({ date, month }: RecordsProps) {
-  const [records, setRecords] = useState<AccountingRecord[]>([]);
-  const [filteredRecords, setFilteredRecords] = useState<AccountingRecord[]>(
-    [],
-  );
-  const [loading, setLoading] = useState(true);
-  const now = new Date();
-
-  useEffect(() => {
-    console.log('2日期變動:', date);
-  }, [date]); // 每次日期變動時觸發
-
-  // 🔹 計算當前月份的時間範圍
-  const startTimestamp = Timestamp.fromDate(
-    startOfMonth(new Date(now.getFullYear(), month, 1)),
-  );
-  const endTimestamp = Timestamp.fromDate(
-    endOfMonth(new Date(now.getFullYear(), month, 1)),
-  );
-
-  // 🔥 查詢當月份的記帳紀錄
-  async function fetchRecords() {
-    try {
-      const q = query(
-        collection(db, 'accounting-records'),
-        where('date', '>=', startTimestamp),
-        where('date', '<=', endTimestamp),
-      );
-      // 使用 onSnapshot 來監聽資料變動
-      const unsubscribe = onSnapshot(q, (querySnapshot) => {
-        const data = querySnapshot.docs.map((doc) => ({
-          id: doc.id,
-          ...doc.data(),
-        })) as AccountingRecord[];
-
-        console.log('data', data);
-
-        setRecords(data);
-        // 根據日期篩選資料
-        if (date) {
-          const filtered = data.filter((record) =>
-            isSameDay(record.date.toDate(), date),
-          );
-          console.log('filteredRecords2', filtered);
-          setFilteredRecords(filtered);
-        }
-      });
-
-      // 清理監聽器
-      return () => unsubscribe();
-    } catch (error) {
-      console.error('Error fetching records:', error);
-    }
-  }
-
-  // 🔥 即時監聽當月份的 Firestore 記錄
-  useEffect(() => {
-    setLoading(true);
-    fetchRecords();
-    setLoading(false);
-  }, [month]); // 每次月份變動時觸發
-
-  // 🔥 處理當日期變動時的篩選
-  useEffect(() => {
-    if (date) {
-      const filteredRecords = records.filter((record) =>
-        isSameDay(record.date.toDate(), date),
-      );
-      console.log('filteredRecords2', filteredRecords);
-      setFilteredRecords(filteredRecords);
-    }
-  }, [date]); // 當日期變動時，根據日期篩選記錄
+  const { filteredRecords, loading } = useAccountingRecords(date, month);
 
   return (
     <div className="rounded-2xl">
